--- api/.flaskenv
@@ -1,2 +1,0 @@
-FLASK_APP=test_frontend.py
-FLASK_ENV=development
