+++ api/Procfile
@@ -1,1 +1,1 @@
-web: gunicorn api.flask_frontend:app
+web: gunicorn api.run:app
