+++ api/app/routes.py
@@ -1,71 +1,13 @@
 import json
 import random
-from flask import Flask, Response, request, redirect, url_for, flash
-from flask_cors import CORS
-from flask_login import current_user, login_user, UserMixin, LoginManager, login_required, logout_user
-from flask_sqlalchemy import SQLAlchemy
-from flask_marshmallow import Marshmallow
-from flask_bcrypt import Bcrypt
 
-app = Flask(__name__)
-CORS(app, supports_credentials=True, origins="https://kraut-und-rueben.herokuapp.com")
-app.config['SQLALCHEMY_DATABASE_URI'] = "sqlite:///user.db"
-app.config["SECRET_KEY"] = "test1234"
-app.config['SESSION_COOKIE_SAMESITE'] = "None"
-app.config['SESSION_COOKIE_SECURE'] = True
-db = SQLAlchemy(app)
-ma = Marshmallow(app)
-bcrypt = Bcrypt(app)
-login_manager = LoginManager(app)
-login_manager.login_view = "login"
+from flask import *
+from flask_login import login_user, logout_user, current_user
 
-
-class User(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(20), unique=True, nullable=False)
-    password = db.Column(db.String(60), nullable=False)
-    recipes = db.relationship("Recipes", backref="owner", lazy=True)
-    ingredients = db.relationship("Ingredients", backref="owner", lazy=True)
-
-    def __repr__(self):
-        return f"{self.username}"
-
-
-class Recipes(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    recipe_name = db.Column(db.String(50), unique=False, nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-    ingredients = db.relationship("Ingredients", backref="recipe_name", lazy=True)
-
-    def __repr__(self):
-        return f"{self.recipe_name}"
-
-
-class Ingredients(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    ingredient = db.Column(db.String(30), unique=False, nullable=False)
-    amount = db.Column(db.Integer, nullable=False)
-    unit = db.Column(db.String(20), nullable=False)
-    recipes_id = db.Column(db.String(50), db.ForeignKey("recipes.recipe_name"), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey("user.id"), nullable=False)
-
-    def __repr__(self):
-        return f"{self.ingredient}, {self.amount},{self.unit}"
-
-
-class RecipeSchema(ma.SQLAlchemyAutoSchema):
-    class Meta:
-        model = Recipes
-
-
-class IngredientSchema(ma.SQLAlchemyAutoSchema):
-    class Meta:
-        model = Ingredients
-
-
-@login_manager.user_loader
-def load_user(user_id):
-    return User.query.get(int(user_id))
+from app import app, bcrypt, db, login_manager
+from app.models.ingredient import Ingredient, IngredientSchema
+from app.models.recipe import Recipe, RecipeSchema
+from app.models.user import User
 
 
 @app.route("/register", methods=["POST"])
@@ -122,7 +64,7 @@ def logout():
 
 
 @app.route("/addingredients", methods=["POST"])
-#@login_required
+# @login_required
 def add_ingredient():
     # username = "steffen"
     # password = "test"
@@ -131,7 +73,7 @@ def add_ingredient():
     # login_user(user)
     get_ingredients = request.get_json()
     recipe_name = get_ingredients["name"]
-    recipe = Recipes(recipe_name=recipe_name, owner=current_user)
+    recipe = Recipe(recipe_name=recipe_name, owner=current_user)
     db.session.add(recipe)
     db.session.commit()
     for element in range(len(get_ingredients["ingredients"])):
@@ -151,16 +93,16 @@ def add_ingredient():
             ingredient = get_ingredients["ingredients"][element]["name"]
             amount = get_ingredients["ingredients"][element]["number"]
             unit = get_ingredients["ingredients"][element]["unit"]
-            current_recipe = Recipes.query.filter_by(recipe_name=recipe_name).first()
-            ingredients = Ingredients(ingredient=ingredient, amount=amount, unit=unit, recipe_name=current_recipe,
-                                      owner=current_user)
+            current_recipe = Recipe.query.filter_by(recipe_name=recipe_name).first()
+            ingredients = Ingredient(ingredient=ingredient, amount=amount, unit=unit, recipe_name=current_recipe,
+                                     owner=current_user)
             db.session.add(ingredients)
             db.session.commit()
     return "Success"
 
 
 @app.route("/randommealplan", methods=["GET"])
-#@login_required
+# @login_required
 def random_recipes():
     # username = "steffen"
     # password = "test"
@@ -168,7 +110,7 @@ def random_recipes():
     # if user and bcrypt.check_password_hash(user.password, password):
     # login_user(user)
     recipe_list = []
-    recipe_list_db = Recipes.query.filter_by(user_id=current_user.id).all()
+    recipe_list_db = Recipe.query.filter_by(user_id=current_user.id).all()
     recipe_schema = RecipeSchema(many=True)
     recipe_list_json = recipe_schema.dump(recipe_list_db)
     for recipe in recipe_list_json:
@@ -189,22 +131,24 @@ def random_recipes():
 
 
 @app.route("/showingredients", methods=["GET"])
-#@login_required
+# @login_required
 def show_ingredients():
     recipe_name = request.form.get("recipe")
-    ingredient_list_db = Ingredients.query.filter_by(recipes_id=recipe_name, user_id=current_user.id).all()
+    ingredient_list_db = Ingredient.query.filter_by(recipes_id=recipe_name, user_id=current_user.id).all()
     ingredient_schema = IngredientSchema(many=True)
     ingredient_list_json = ingredient_schema.dump(ingredient_list_db)
     return Response(json.dumps(ingredient_list_json), mimetype="application/json")
 
+
 @app.route("/recipeoverview", methods=["GET"])
-#@login_required
+# @login_required
 def recipe_overview():
-    recipe_list_db = Recipes.query.filter_by(user_id=current_user.id).all()
+    recipe_list_db = Recipe.query.filter_by(user_id=current_user.id).all()
     recipe_schema = RecipeSchema(many=True)
     recipe_list = recipe_schema.dump(recipe_list_db)
     return Response(json.dumps(recipe_list), mimetype="application/json")
 
 
-if __name__ == "__main__":
-    app.run(threaded=True, port=5000)
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
