+++ api/config.py
@@ -1,0 +1,4 @@
+SQLALCHEMY_DATABASE_URI = "sqlite:///user.db"
+SECRET_KEY = "test1234"
+SESSION_COOKIE_SAMESITE = "None"
+SESSION_COOKIE_SECURE = True
