+++ api/run.py
@@ -1,0 +1,3 @@
+from app import app
+
+app.run(threaded=True, port=5000)
