--- api/templates/add_ingredients.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<body>
-<form method="Post">
-    <div><label>Rezeptname: <input type="text" name ="recipe_name"> </label>
-        <label> Zutat: <input type="text" name="ingredient"></label>
-        <label> Menge: <input type="text" name="amount"></label>
-        <label> Einheit: <input type="text" name="unit"></label>
-    </div>
-    <input type="submit" name="add_ingredient" value="Füge die Zutat zum Rezept hinzu">
-    <input type="submit" name="redirect_index" value="Zurück zur Startseite.">
-</form>
-</body>
-</html>
