--- api/templates/generate_random_recipes.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<body>
-<form method="Post">
-    <input type="submit" name="generate_meal_plan" value="Generiere meinen Wochenplan">
-    <input type="submit" name="redirect_index" value="Zurück zur Startseite">
-</form>
-</body>
-</html>
