--- api/templates/login.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Benutzerauswahl</title>
-</head>
-<body>
-<form method="Post">
-    <div><label>Bitte gib deinen Benutzernamen ein: <input type="text" name ="username"></label></div>
-    <div><label>Bitte gib dein Passwort ein: <input type="password" name ="password"></label></div>
-    <input type="submit" name="random_meal_plan" value="Zeig mir meine Rezepte!">
-    <input type="submit" name="redirect_index" value="Zurück zur Startseite">
-</form>
-</body>
-</html>
