--- api/templates/register.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-<form method="Post">
-    <div><label>Bitte gib deinen gewünschten Benutzernamen ein: <input type="text" name ="username"></label></div>
-    <div><label>Bitte gib dein Passwort ein: <input type="password" name ="password"></label></div>
-    <input type="submit" name="register_user" value="Erstelle den Account!">
-    <input type="submit" name="redirect_index" value="Zurück zur Startseite!">
-</body>
-</html>
