--- api/templates/show_ingredients.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<form method="Post">
-    <div><label>Rezeptname: <input type="text" name ="recipe"></label></div>
-    <input type="submit" name="submit" value="Rezept hinzufügen">
-    <input type="submit" name="redirect_index" value="Zurück zur Startseite">
-</form>
